refactor: move `getRank` function to its own file
